requirements.txt
pytorch
flask
langchain
langchain-community
langchain-huggingface
unstructured[pdf]
faiss-cpu
sentence-transformers
transformers
python-dotenv
pytesseract